style(front): restructures frontend code, adds aliases, and refactors imports
closes #428
