+++ front/src/lib/components/Loader.svelte
