+++ front/src/lib/components/NavToggle.svelte
