+++ front/src/lib/components/icons/Catchafire.svelte
