+++ front/src/lib/layout/Footer.svelte
