+++ front/src/lib/layout/NavBar.svelte
