+++ front/src/lib/layout/NavHome.svelte
