+++ front/src/lib/layout/ScrollTop.svelte
@@ -1,6 +1,6 @@
 <script>
   import IconUp from "~icons/icon-park-outline/up-c";
-  import { scrollTop } from "../_utils";
+  import { scrollTop } from "@utils";
 </script>
 
 <a href="#header" on:click={scrollTop} class="scroll-top">
