+++ front/src/lib/layout/ThemeToggleIcon.svelte
