+++ front/src/lib/styles/pages/not-found.css
