+++ front/src/routes/+error.svelte
@@ -1,7 +1,7 @@
 <script>
   import * as Sentry from "@sentry/sveltekit";
   import { PUBLIC_ENV, PUBLIC_SENTRY_DSN } from "$env/static/public";
-  import "../styles/pages/not-found.css";
+  import "@styles/pages/not-found.css";
   import { page } from "$app/stores";
 
   Sentry.captureMessage("Page Not Found", {
