+++ front/src/routes/+layout.svelte
@@ -4,12 +4,12 @@
   import { PUBLIC_ENV } from "$env/static/public";
   import { PlausibleAnalytics } from "@accuser/svelte-plausible-analytics";
   import { onMount } from "svelte";
-  import { scrollTop } from "../_utils/index.js";
-  import Head from "../components/Head.svelte";
-  import Flourish from "../layout/Flourish.svelte";
-  import Footer from "../layout/Footer.svelte";
-  import Header from "../layout/Header.svelte";
-  import "../styles/global.css";
+  import { scrollTop } from "@utils";
+  import Head from "@components/Head.svelte";
+  import Flourish from "@layout/Flourish.svelte";
+  import Footer from "@layout/Footer.svelte";
+  import Header from "@layout/Header.svelte";
+  import "@styles/global.css";
 
   export let data;
   const { navHeading, navItems, copyright, copyleft, seo } = data;
